Add another step for 6MB
Changes the steps for the 6MB from 8 > 6(2+4) > 1 to 8 > 6 (2+4) > 2 > 1

Consider a system configured with a 4MB board before the GGF,
Before this change, it would result in a total of 7MB RAM

With the old code the following steps would happen
1. GGF appears as an 8MB board and is told to shut-up
2. GGF attempts to configure as a 6MB board by offering first a 2M then a
4MB block
  1. The 2MB part of the 6M option will be accepted
  2. The 4MB part will be denied because there's no space for it
  3. The GGF then falls back to it's 1MB block which is accepted
3: With the process finished the system now has 7MB of Fast RAM

With the new code step 2.3 becomes an offering of 2MB - Thus giving 8MB
Fast Ram.

Credit to @cdhooper for suggesting this improvement.